konsoliderer brevmottaker komponenter
